# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RealmForge Control API",
    "description": "Governance kernel REST API — session, command, audit, snapshot, rollback, boards, and more.",
    "license": {
      "name": "UNLICENSED",
      "identifier": "UNLICENSED"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/boards/plans": {
      "get": {
        "tags": [
          "boards"
        ],
        "summary": "List board plans with optional status filter and pagination.",
        "operationId": "list_board_plans",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by board state. One of: draft, in_review, approved, in_progress, completed, blocked, archived.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plans listed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardPlanListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "boards"
        ],
        "summary": "Create a new board plan for a work path.",
        "operationId": "create_board_plan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBoardPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Plan created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardPlanResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/boards/plans/{id}/approve": {
      "post": {
        "tags": [
          "boards"
        ],
        "summary": "Approve a board plan submitted for review.",
        "operationId": "approve_board_plan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Board plan ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Plan approved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardPlanResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Plan not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/boards/plans/{id}/reject": {
      "post": {
        "tags": [
          "boards"
        ],
        "summary": "Reject a board plan and return it to draft.",
        "operationId": "reject_board_plan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Board plan ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Plan rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardPlanResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Plan not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/boards/plans/{id}/submit": {
      "post": {
        "tags": [
          "boards"
        ],
        "summary": "Submit a draft board plan for approval.",
        "operationId": "submit_board_plan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Board plan ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plan submitted for review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardPlanResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid plan ID or state transition",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Plan not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/boards/releases": {
      "post": {
        "tags": [
          "boards"
        ],
        "summary": "Submit a release command from an approved board plan.",
        "operationId": "submit_board_release",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardReleaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Release command submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardReleaseResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BoardApprovalRequest": {
        "type": "object",
        "required": [
          "approver"
        ],
        "properties": {
          "approver": {
            "type": "string"
          },
          "comment": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "BoardPlanListResponse": {
        "type": "object",
        "required": [
          "plans",
          "total"
        ],
        "properties": {
          "plans": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BoardPlanResponse"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "BoardPlanResponse": {
        "type": "object",
        "required": [
          "id",
          "title",
          "status",
          "work_path_refs",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "work_path_refs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "BoardReleaseRequest": {
        "type": "object",
        "required": [
          "plan_id",
          "bundle_ref",
          "approval_ref"
        ],
        "properties": {
          "approval_ref": {
            "type": "string"
          },
          "bundle_ref": {
            "type": "string"
          },
          "plan_id": {
            "type": "string"
          }
        }
      },
      "BoardReleaseResponse": {
        "type": "object",
        "required": [
          "command_id",
          "status",
          "plan_id"
        ],
        "properties": {
          "command_id": {
            "type": "string"
          },
          "plan_id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "CreateBoardPlanRequest": {
        "type": "object",
        "required": [
          "title"
        ],
        "properties": {
          "title": {
            "type": "string",
            "example": "Login Vertical — Phase 9"
          },
          "work_path_refs": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 7807 Problem Details response, extended with a machine-readable error code\nand trace ID per the RealmForge API error standard.",
        "required": [
          "code",
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Machine-readable SCREAMING_SNAKE_CASE error code. Stable across versions."
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "trace_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Correlation ID matching the X-Request-Id response header."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "session_token"
      }
    }
  },
...
}